+++ packages/net-kit/lib/src/manager/adapter/platform_http_adapter.dart
@@ -1,0 +1,3 @@
+export 'platform_http_adapter_stub.dart'
+    if (dart.library.io) 'platform_http_adapter_io.dart'
+    if (dart.library.html) 'platform_http_adapter_web.dart';