+++ packages/net-kit/lib/src/manager/adapter/platform_http_adapter_stub.dart
@@ -1,0 +1,6 @@
+import 'i_http_adapter.dart';
+
+/// The class implements the IHttpAdapter interface
+IHttpAdapter createPlatformAdapter() {
+  throw UnsupportedError('No adapter available for this platform.');
+}