+++ packages/net-kit/lib/src/manager/net_kit_manager.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:dio/dio.dart';
 
 import '../enum/http_status_codes.dart';
@@ -13,9 +11,7 @@ import '../utility/converter.dart';
 import '../utility/logger/i_net_kit_logger.dart';
 import '../utility/logger/void_logger.dart';
 import '../utility/typedef/request_type_def.dart';
-import 'adapter/io_http_adapter.dart'
-if (dart.library.io) 'adapter/io_http_adapter.dart'
-if (dart.library.html) 'adapter/web_http_adapter.dart';
+import 'adapter/platform_http_adapter.dart';
 import 'error/api_exception.dart';
 import 'i_net_kit_manager.dart';
 import 'params/net_kit_error_params.dart';
@@ -27,8 +23,7 @@ import 'token/token_manager.dart';
 part 'interceptors/error_handling_interceptor.dart';
 part 'mixin/error_handling_mixin.dart';
 part 'mixin/request_manager_mixin.dart';
-part 'mixin/token_manager_mixin.dart';.dart';
-
+part 'mixin/token_manager_mixin.dart';
 part 'mixin/upload_manager_mixin.dart';
 
 /// The NetKitManager class is a network manager that extends DioMixin and
@@ -77,9 +72,6 @@ class NetKitManager extends INetKitManager
     /// The key for the access token in the headers
     String accessTokenHeaderKey = 'Authorization',
 
-    /// The key for the refresh token in the headers
-    String refreshTokenHeaderKey = 'Refresh-Token',
-
     /// The key for the access token in the body
     /// Used for automatic token refresh
     String accessTokenBodyKey = 'accessToken',
@@ -117,7 +109,6 @@ class NetKitManager extends INetKitManager
       logger: logger,
       internetStatusStream: internetStatusStream,
       accessTokenHeaderKey: accessTokenHeaderKey,
-      refreshTokenHeaderKey: refreshTokenHeaderKey,
       accessTokenBodyKey: accessTokenBodyKey,
       refreshTokenBodyKey: refreshTokenBodyKey,
       dataKey: dataKey,
@@ -449,7 +440,6 @@ class NetKitManager extends INetKitManager
     required bool logInterceptorEnabled,
     required bool testMode,
     required String accessTokenHeaderKey,
-    required String refreshTokenHeaderKey,
     required String accessTokenBodyKey,
     required String refreshTokenBodyKey,
     required INetKitLogger logger,
@@ -480,7 +470,6 @@ class NetKitManager extends INetKitManager
       testMode: testMode,
       logInterceptorEnabled: logInterceptorEnabled,
       accessTokenHeaderKey: accessTokenHeaderKey,
-      refreshTokenHeaderKey: refreshTokenHeaderKey,
       accessTokenBodyKey: accessTokenBodyKey,
       refreshTokenBodyKey: refreshTokenBodyKey,
       refreshToken: RefreshTokenParams(
@@ -498,7 +487,7 @@ class NetKitManager extends INetKitManager
     );
 
     /// Set up the http client adapter
-    httpClientAdapter = clientAdapter ?? HttpAdapter().getAdapter();
+    httpClientAdapter = clientAdapter ?? createPlatformAdapter().getAdapter();
 
     /// If test mode is enabled, use devBaseUrl
     parameters.testMode
@@ -517,9 +506,9 @@ class NetKitManager extends INetKitManager
       refreshTokenPath: parameters.refreshToken.refreshTokenPath,
       requestQueue: RequestQueue(),
       tokenManager: TokenManager(
-        addBearerToken: setAccessToken,
-        addRefreshToken: setRefreshToken,
-        refreshTokenRequest: _refreshTokenRequest,
+        addBearerToken: _setAccessToken,
+        addRefreshToken: _setRefreshToken,
+        refreshTokenRequest: _requestNewTokens,
         retryRequest: _retryRequest,
         onTokensUpdated: _onTokensUpdated,
         logger: _logger,
@@ -555,11 +544,24 @@ class NetKitManager extends INetKitManager
     parameters.internetStatusSubscription?.cancel();
   }
 
-  /// Method to get the access token from the headers.
-  String getRefreshToken() {
-    final token = baseOptions.headers[parameters.refreshTokenHeaderKey];
-    if (token == null) return '';
-    return token is String ? token : '';
+  @override
+  void setAccessToken(String? token) {
+    _setAccessToken(token);
+  }
+
+  @override
+  void setRefreshToken(String? token) {
+    _setRefreshToken(token);
+  }
+
+  @override
+  void removeRefreshToken() {
+    _removeRefreshToken();
+  }
+
+  @override
+  void removeAccessToken() {
+    _removeAccessToken();
   }
 
   /// Method to be called when the tokens are updated.
@@ -571,11 +573,14 @@ class NetKitManager extends INetKitManager
     if (onTokenRefreshed != null) {
       onTokenRefreshed!.call(authToken);
     }
+    _setAccessToken(authToken.accessToken);
+    _setRefreshToken(authToken.refreshToken);
   }
 
-  Future<AuthTokenModel> _refreshTokenRequest(String refreshTokenPath) async {
+  /// Method to send a request to the server to refresh the access token.
+  Future<AuthTokenModel> _requestNewTokens(String refreshTokenPath) async {
     // Remove the access token before refreshing
-    removeAccessToken();
+    _removeAccessToken();
 
     final refreshResponse = await request<dynamic>(
       refreshTokenPath,
@@ -585,7 +590,7 @@ class NetKitManager extends INetKitManager
       ),
       data: parameters.refreshToken.body ??
           {
-            parameters.refreshTokenBodyKey: getRefreshToken(),
+            parameters.refreshTokenBodyKey: _refreshToken,
           },
     );
     return extractTokens(response: refreshResponse);