+++ packages/net-kit/test/src/manager/manager/net_kit_manager_test.dart
@@ -23,6 +23,7 @@ void main() {
         baseUrl: 'https://<TEST-API>.com',
         internetStatusStream: internetStatusController.stream,
       );
+
       netKitManagerWithCustomDataKey = NetKitManager(
         baseUrl: 'https://<TEST-API>.com',
         internetStatusStream: internetStatusController.stream,
@@ -35,7 +36,6 @@ void main() {
         accessTokenBodyKey: 'access_token',
         refreshTokenBodyKey: 'refresh_token',
         accessTokenHeaderKey: 'custom_access_token_header',
-        refreshTokenHeaderKey: 'custom_refresh_token_header',
       );
 
       netKitManagerWithCustomKeysAndDataKey = NetKitManager(
@@ -44,7 +44,6 @@ void main() {
         accessTokenBodyKey: 'access_token',
         refreshTokenBodyKey: 'refresh_token',
         accessTokenHeaderKey: 'custom_access_token_header',
-        refreshTokenHeaderKey: 'custom_refresh_token_header',
         dataKey: 'customData',
       );
     });
@@ -99,8 +98,7 @@ void main() {
       });
 
       test('should return null tokens when tokens are missing', () {
-        final response =
-            Response<dynamic>(requestOptions: RequestOptions(path: '/test'));
+        final response = Response<dynamic>(requestOptions: RequestOptions(path: '/test'));
 
         final tokens = netKitManager.extractTokens(response: response);
 
@@ -108,8 +106,7 @@ void main() {
         expect(tokens.refreshToken, isNull);
       });
 
-      test('should return null access token when only refresh token is present',
-          () {
+      test('should return null access token when only refresh token is present', () {
         final response = Response<dynamic>(
           requestOptions: RequestOptions(path: '/test'),
           data: <String, dynamic>{'refreshToken': 'refresh-token-value'},
@@ -121,8 +118,7 @@ void main() {
         expect(tokens.refreshToken, 'refresh-token-value');
       });
 
-      test('should return null refresh token when only access token is present',
-          () {
+      test('should return null refresh token when only access token is present', () {
         final response = Response<dynamic>(
           requestOptions: RequestOptions(path: '/test'),
           data: <String, dynamic>{'accessToken': 'access-token-value'},
@@ -145,8 +141,7 @@ void main() {
           },
         );
 
-        final tokens =
-            netKitManagerWithCustomKeys.extractTokens(response: response);
+        final tokens = netKitManagerWithCustomKeys.extractTokens(response: response);
 
         expect(tokens.accessToken, 'access-token-value');
         expect(tokens.refreshToken, 'refresh-token-value');
@@ -209,8 +204,7 @@ void main() {
         expect(tokens.refreshToken, isNull);
       });
 
-      test('should return null tokens when response has an error status code',
-          () {
+      test('should return null tokens when response has an error status code', () {
         final response = Response<dynamic>(
           requestOptions: RequestOptions(path: '/test'),
           statusCode: 500, // Internal server error
@@ -264,19 +258,16 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, 'access-token-value');
           expect(tokens.refreshToken, 'refresh-token-value');
         });
 
         test('should return null tokens when tokens are missing', () {
-          final response =
-              Response<dynamic>(requestOptions: RequestOptions(path: '/test'));
+          final response = Response<dynamic>(requestOptions: RequestOptions(path: '/test'));
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, isNull);
           expect(tokens.refreshToken, isNull);
@@ -294,8 +285,7 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, isNull);
           expect(tokens.refreshToken, 'refresh-token-value');
@@ -313,8 +303,7 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, 'access-token-value');
           expect(tokens.refreshToken, isNull);
@@ -374,8 +363,7 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, isNull);
           expect(tokens.refreshToken, isNull);
@@ -392,8 +380,7 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, '');
           expect(tokens.refreshToken, '');
@@ -408,15 +395,13 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, isNull);
           expect(tokens.refreshToken, isNull);
         });
 
-        test('should return null tokens when response has an error status code',
-            () {
+        test('should return null tokens when response has an error status code', () {
           final response = Response<dynamic>(
             requestOptions: RequestOptions(path: '/test'),
             statusCode: 500, // Internal server error
@@ -428,15 +413,13 @@ void main() {
             },
           );
 
-          final tokens =
-              netKitManagerWithCustomDataKey.extractTokens(response: response);
+          final tokens = netKitManagerWithCustomDataKey.extractTokens(response: response);
 
           expect(tokens.accessToken, isNull);
           expect(tokens.refreshToken, isNull);
         });
 
-        test('should extract tokens correctly in concurrent requests',
-            () async {
+        test('should extract tokens correctly in concurrent requests', () async {
           final response = Response<dynamic>(
             requestOptions: RequestOptions(path: '/test'),
             data: {